Refactored map_product into a single file
Repeating code is bad and the functions have already diverged
slightly...
